// src/app/model/EstimatedItemsModel.ts
export class EstimatedItemsModel {
    EstimateItemId: number;
    EstimateId: number;
    OrganizationId: number;
    OrganizationName: string;
    AdjustedItemName: string;
    AdjustLTypeId: number;
    AdjustmentType: string;
    AdjustmentValue: number;
    AdjustedUnitCost: number;
    Qty: number;
    TotalCost: number;
    CostBreakDownId: number;
    CostBreakDownHeading: string;
    CostBreakDownName: string;
    ItemId: string;
    ItemCode: string;
    Note: string;
    Remarks: string;
    Category: string;
    SubCategory: string;
    Unit: string;
    UnitPrice: string;

}